Better support of Sherpa RootAnalysis plugin.
Can run several parallel analyses at the same time now.

+++ SelectorCommon.h
@@ -340,7 +340,7 @@ class SelectorCommon : public TSelector
     double pdf_warning_thresh;
 
     // exact reweighting
-    int event_trials;
+    double event_trials;
 
   protected:
     double get_alphas(double mur);
